requirements.txt
typer[all]>=0.9.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0
